Move reduction mocks into separate MockObjects file
