+++ qt/scientific_interfaces/Indirect/test/Reduction/ISISEnergyTransferPresenterTest.h
@@ -9,14 +9,13 @@
 #include <cxxtest/TestSuite.h>
 #include <gmock/gmock.h>
 
+#include "MockObjects.h"
 #include "Reduction/ISISEnergyTransferPresenter.h"
-#include "Reduction/IndirectDataReduction.h"
 
 #include "MantidAPI/AlgorithmProperties.h"
 #include "MantidKernel/WarningSuppressions.h"
 
 using namespace MantidQt::CustomInterfaces;
-using MantidQt::API::BatchAlgorithmRunner;
 using namespace testing;
 
 namespace {
@@ -28,88 +27,6 @@ std::unique_ptr<Mantid::API::AlgorithmRuntimeProps> defaultGroupingProps() {
 }
 
 } // namespace
-
-GNU_DIAG_OFF_SUGGEST_OVERRIDE
-
-class MockIETModel : public IIETModel {
-
-public:
-  MOCK_METHOD2(setInstrumentProperties, void(IAlgorithmRuntimeProps &properties, InstrumentData const &instData));
-
-  MOCK_METHOD1(validateRunData, std::vector<std::string>(IETRunData const &runData));
-  MOCK_METHOD1(validatePlotData, std::vector<std::string>(IETPlotData const &plotData));
-
-  MOCK_METHOD3(runIETAlgorithm, std::string(MantidQt::API::BatchAlgorithmRunner *batchAlgoRunner,
-                                            InstrumentData const &instData, IETRunData &runParams));
-  MOCK_CONST_METHOD2(plotRawAlgorithmQueue,
-                     std::deque<MantidQt::API::IConfiguredAlgorithm_sptr>(InstrumentData const &instData,
-                                                                          IETPlotData const &plotData));
-
-  MOCK_METHOD2(saveWorkspace, void(std::string const &workspaceName, IETSaveData const &saveData));
-
-  MOCK_METHOD4(createGroupingWorkspace, void(std::string const &instrumentName, std::string const &analyser,
-                                             std::string const &customGrouping, std::string const &outputName));
-
-  MOCK_METHOD1(loadDetailedBalance, double(std::string const &filename));
-
-  MOCK_METHOD4(groupWorkspaces, std::vector<std::string>(std::string const &groupName, std::string const &instrument,
-                                                         std::string const &groupOption, bool const shouldGroup));
-};
-
-class MockIETView : public IIETView {
-public:
-  MOCK_METHOD1(subscribePresenter, void(IIETPresenter *presenter));
-
-  MOCK_CONST_METHOD0(getRunData, IETRunData());
-  MOCK_CONST_METHOD0(getPlotData, IETPlotData());
-  MOCK_CONST_METHOD0(getSaveData, IETSaveData());
-  MOCK_CONST_METHOD0(getGroupOutputOption, std::string());
-  MOCK_CONST_METHOD0(getPlotOptionsView, OutputPlotOptionsView *());
-  MOCK_CONST_METHOD0(getGroupOutputCheckbox, bool());
-  MOCK_CONST_METHOD0(getFirstFilename, std::string());
-  MOCK_CONST_METHOD0(isRunFilesValid, bool());
-  MOCK_CONST_METHOD1(validateCalibrationFileType, void(UserInputValidator &uiv));
-  MOCK_CONST_METHOD1(validateRebinString, void(UserInputValidator &uiv));
-  MOCK_CONST_METHOD2(validateGroupingProperties,
-                     std::optional<std::string>(std::size_t const &spectraMin, std::size_t const &spectraMax));
-  MOCK_CONST_METHOD0(showRebinWidthPrompt, bool());
-  MOCK_CONST_METHOD3(showSaveCustomGroupingDialog,
-                     void(std::string const &customGroupingOutput, std::string const &defaultGroupingFilename,
-                          std::string const &saveDirectory));
-  MOCK_CONST_METHOD1(displayWarning, void(std::string const &message));
-  MOCK_METHOD1(setCalibVisible, void(bool visible));
-  MOCK_METHOD1(setEfixedVisible, void(bool visible));
-  MOCK_METHOD1(setBackgroundSectionVisible, void(bool visible));
-  MOCK_METHOD1(setPlotTimeSectionVisible, void(bool visible));
-  MOCK_METHOD1(setAnalysisSectionVisible, void(bool visible));
-  MOCK_METHOD1(setPlottingOptionsVisible, void(bool visible));
-  MOCK_METHOD1(setAclimaxSaveVisible, void(bool visible));
-  MOCK_METHOD1(setSPEVisible, void(bool visible));
-  MOCK_METHOD1(setFoldMultipleFramesVisible, void(bool visible));
-  MOCK_METHOD1(setOutputInCm1Visible, void(bool visible));
-  MOCK_METHOD1(setGroupOutputCheckBoxVisible, void(bool visible));
-  MOCK_METHOD1(setGroupOutputDropdownVisible, void(bool visible));
-  MOCK_METHOD1(setInstrumentDefault, void(InstrumentData const &instrumentDetails));
-  MOCK_METHOD1(setDetailedBalance, void(double detailedBalance));
-  MOCK_METHOD1(setRunFilesEnabled, void(bool enable));
-  MOCK_METHOD1(setSingleRebin, void(bool enable));
-  MOCK_METHOD1(setMultipleRebin, void(bool enable));
-  MOCK_METHOD1(setSaveEnabled, void(bool enable));
-  MOCK_METHOD1(setPlotTimeIsPlotting, void(bool plotting));
-  MOCK_METHOD2(setFileExtensionsByName, void(QStringList calibrationFbSuffixes, QStringList calibrationWSSuffixes));
-  MOCK_METHOD1(setOutputWorkspaces, void(std::vector<std::string> const &outputWorkspaces));
-  MOCK_METHOD1(setRunButtonText, void(std::string const &runText));
-  MOCK_METHOD1(setEnableOutputOptions, void(bool const enable));
-};
-
-class MockIndirectDataReduction : public IIndirectDataReduction {
-  MOCK_METHOD0(instrumentWorkspace, Mantid::API::MatrixWorkspace_sptr());
-
-  MOCK_CONST_METHOD0(getInstrumentConfiguration, MantidQt::MantidWidgets::IndirectInstrumentConfig *());
-  MOCK_METHOD0(getInstrumentDetails, QMap<QString, QString>());
-
-  MOCK_METHOD1(showAnalyserAndReflectionOptions, void(bool visible));
-};
 
 class IETPresenterTest : public CxxTest::TestSuite {
 public:
